# src/main/resources/static/api.yaml
openapi: "3.0.0"
info:
  version: "1"
  title: darfichraus.de API
servers:
  - url: https://api.darfichraus.de
    description: production instance
  - url: https://api.dev.crimsy.tech
    description: development instance
security:
  - ApiKeyAuth: []
...
paths:
...
  /subscriptions:
    post:
      tags:
        - subscription
      operationId: addSubscription
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscription'
      responses:
        '204':
          description: successfully added additional information
  /subscriptions/{id}:
    delete:
      tags:
        - subscription
      operationId: deleteSubscription
      summary: deletes a Subscription
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '204':
          description: successfully deleted Subscription
...
  /situation-advisor/situation-types/{id}:
    get:
      tags:
        - situationTypes_Public
      operationId: getSituationType
      summary: Get specific Situation Type
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: return a specific Situation Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationType'
...
  /situation-advisor/situations:
    get:
      tags:
        - situations_Public
      security:
        - BearerAuth: []
      operationId: getAllSituations
      summary: Retrieves a list with all situations
      responses:
        '200':
          description: List of all situations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Situation'
  /situation-advisor/situations/{id}:
    get:
      tags:
        - situations_Public
      operationId: getSituation
      summary: Get specific Situation
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: return a specific Situation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Situation'
...
  /situation-advisor/situations-message-types:
    get:
      tags:
        - situationMessageTypes_Public
      security:
        - BearerAuth: []
      operationId: getAllSituationMessageTypes
      summary: Retrieves a list with all Situation Message Types
      responses:
        '200':
          description: List of all Situation Message Types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SituationMessageType'
  /situation-advisor/situation-message-types/{id}:
    get:
      tags:
        - situationMessageTypes_Public
      operationId: getSituationMessageType
      summary: Get specific Situation Message Type
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: return a specific Situation Message Type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationMessageType'
...
  /situation-advisor/situations-messages:
    get:
      tags:
        - situationMessages_Public
      security:
        - BearerAuth: []
      operationId: getAllSituationMessages
      summary: Retrieves a list with all Situation Messages
      parameters:
        - in: query
          name: lastRequest
          description: "Retrieve only the delta of information since last requst"
          required: false
          example: "2020-05-04T16:01:37.933Z"
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: List of all Situation Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SituationMessage'
  /situation-advisor/situation-messages/{id}:
    get:
      tags:
        - situationMessages_Public
      operationId: getSituationMessage
      summary: Get specific Situation Message
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: return a specific Situation Message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationMessage'
...
  /situation-advisor/situations-references:
    get:
      tags:
        - situationReferences_Public
      security:
        - BearerAuth: []
      operationId: getAllSituationReferences
      summary: Retrieves a list with all Situation References
      responses:
        '200':
          description: List of all Situation References
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SituationReference'
  /situation-advisor/situation-references/{id}:
    get:
      tags:
        - situationReferences_Public
      operationId: getSituationReference
      summary: Get specific Situation Reference
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: return a specific Situation Reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SituationReference'
...
components:
  schemas:
...
    Areal:
      description: 'A restriction can be city-based (zip), county based (Landkreis), for the whole state (Bayern) or countrywide (country)'
      type: string
      enum:
        - ZIP
        - COUNTY
        - STATE
        - COUNTRY
...
    RestrictionType:
      description: 'A restriction can affect one of these  areas of public life'
      type: string
      enum:
        - GENERAL_INFORMATION
        - PUBLIC_TRANSPORTATION
        - EVENTS_AND_ASSEMBLIES
        - GASTRONOMY
        - PUBLIC_PLACES
        - RETAIL
        - CURFEW
...
    Subscription:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          required:
            - email
            - areal
            - arealIdentifier
            - contactAllowed
          properties:
            email:
              type: string
              example: 'a@b.com'
            areal:
              $ref: '#/components/schemas/Areal'
            arealIdentifier:
              type: string
            types:
              type: array
              items:
                $ref: '#/components/schemas/RestrictionType'
    SituationType:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - required:
            - name
            - description
            - severity
        - type: object
          description: "This object allows to categorize Situations into high level crysis."
          properties:
            name:
              type: string
              description: "Classifier."
              example: "Pandemie"
            description:
              type: string
              example: "Ausbruch einer Krankheit. Betrifft den gesamten Globus"
            severity:
              $ref: '#/components/schemas/SituationSeverity'
            media:
              type: string
              example: "virus.png"
    Situation:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          description: "A Situation is a current crysis."
          required:
            - name
            - startDate
            - severity
          properties:
            name:
              type: string
              description: "Describes the situation; like a title"
              example: "Corona/Covid-19"
            startDate:
              type: string
              format: date
              description: "Start of the situation/crysis"
              example: "2020-01-01"
            endDate:
              type: string
              format: date
              nullable: true
              description: "End of situation/crysis"
              example: "2020-12-31"
            severity:
              $ref: '#/components/schemas/SituationSeverity'
            active:
              type: boolean
              default: true
              description: "Toggles the afffectness of a situation/crysis. False will not trigger any action to a system/user (e. g. push notification)"
    SituationMessageType:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          description: "Classifies a SituationMessage into a type. Has impact to icon, coloring, severity of messages their handling"
          required:
            - name
            - severity
            - color
          properties:
            name:
              type: string
              description: "Speakable type of Type"
              example: "Restrictive Instruktion"
            media:
              type: string
              description: "Icon"
              example: "exclamation-mark.svg"
            severity:
              $ref: '#/components/schemas/SituationSeverity'
            color:
              type: string
              description: "Provides the possibility to annotate messages/headers with MessageType-based color"
              example: "orange"
    SituationMessage:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          description: "Represents an instruction or an information to recipients or systems."
          required:
            - situationId
            - messageTypeId
            - title
            - message
            - abstract
            - severity
          properties:
            situationId:
              type: string
              description: "Sorts the message to a situation/crysis"
              example: "5eaf148a0d077d75108087d3"
            messageTypeId:
              type: string
              description: "Classifies the Message as type of e.g. Restrictive-Instruction; inherits color schema, severity, etc."
              example: "5eaf17692c1f4e9671ec7111"
            title:
              type: string
              description: "Quick information/abstract about the message"
              example: "Wearing a face mask is mandatory!"
            message:
              type: string
              description: "Whole information. Fully formatable."
              example: "By start of April, 20th 2020 you have to wear a face mask in supermarkets and all shop-alike stores."
            abstract:
              type: string
              description: "Short text (excerpt) from the whole message"
              example: "Please wear a face mask in shopping stores"
            icon:
              type: string
              description: "Icon"
              example: "mask.svg"
            severity:
              $ref: '#/components/schemas/SituationSeverity'
            documents:
              type: array
              items:
                $ref: '#/components/schemas/SituationReference'
            areaOfEffect:
              type: array
              description: "A list of all GeoJson-Ids, where the Situation is taking place/is affected"
              items:
                type: string
            status:
              $ref: '#/components/schemas/SituationMessageStatus'
            version:
              type: number
              description: "Every change increments the version. Full history possible"
    SituationReference:
      allOf:
        - $ref: '#/components/schemas/BaseEntity'
        - type: object
          description: "References to source of information"
          properties:
            title:
              type: string
              description: "Allows to describe the source/reference"
              example: "How to wear a face mask?"
            type:
              $ref: '#/components/schemas/SituationReferenceType'
            classifier:
              $ref: '#/components/schemas/SituationReferenceClassifier'
            source:
              type: string
              description: "points to the source"
              example: "http://darfichraus.de/mask-mandatory.pdf"
    SituationReferenceType:
      description: "Type of reference to build type-based-action"
      type: string
      enum:
        - URL
        - FILE
        - PICTURE
        - VIDEO
        - VOICE
        - OTHER
    SituationReferenceClassifier:
      description: "Classifies the reference as type of"
      type: string
      enum:
        - ATTACHMENT
        - SOURCE
        - OTHER
    SituationMessageStatus:
      description: "Get information about the status"
      type: string
      enum:
        - DRAFT
        - APPROVED
        - REVOKED
        - OTHER
    SituationSeverity:
      description: "Describes the severity of a subject"
      type: number
      enum:
        - 1
        - 2
        - 3
        - 4
        - 5
    BaseEntity:
      type: object
      discriminator:
        propertyName: Entity
      properties:
        id:
          type: string
        modified:
          type: string
          format: date-time
          description: 'when this entity has been last modified'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: 'API-KEY'
      description: 'Api-Key needed to authenticate requests'
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT